backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
pydicom
Pillow
numpy

